+++ engine/src/Options/Interfaces/Particles/Move/IMoveAngle.ts
@@ -1,7 +1,9 @@
+import type { RangeValue } from "../../../../Types";
+
 /**
  * @category Options
  */
 export interface IMoveAngle {
-    offset: number;
-    value: number;
+    offset: RangeValue;
+    value: RangeValue;
 }