+++ engine/src/Options/Interfaces/Particles/Move/IMoveGravity.ts
@@ -1,6 +1,8 @@
+import type { RangeValue } from "../../../../Types";
+
 export interface IMoveGravity {
-    acceleration: number;
+    acceleration: RangeValue;
     enable: boolean;
     inverse: boolean;
-    maxSpeed: number;
+    maxSpeed: RangeValue;
 }