+++ engine/src/Options/Interfaces/Particles/Repulse/IRepulse.ts
@@ -1,4 +1,5 @@
 import type { IValueWithRandom } from "../../IValueWithRandom";
+import type { RangeValue } from "../../../../Types";
 
 /**
  * @category Options
@@ -6,8 +7,8 @@ import type { IValueWithRandom } from "../../IValueWithRandom";
  */
 export interface IRepulse extends IValueWithRandom {
     enabled: boolean;
-    distance: number;
-    duration: number;
-    factor: number;
-    speed: number;
+    distance: RangeValue;
+    duration: RangeValue;
+    factor: RangeValue;
+    speed: RangeValue;
 }