+++ engine/src/Plugins/Absorbers/Options/Interfaces/IAbsorber.ts
@@ -1,4 +1,4 @@
-import type { IColor, ICoordinates } from "../../../../Core";
+import type { IColor, IRangedCoordinates } from "../../../../Core";
 import type { IAbsorberSize } from "./IAbsorberSize";
 import { RecursivePartial } from "../../../../Types";
 
@@ -26,7 +26,7 @@ export interface IAbsorber {
     /**
      * Absorber position, percent values calculated on canvas size
      */
-    position?: RecursivePartial<ICoordinates>;
+    position?: RecursivePartial<IRangedCoordinates>;
 
     /**
      * Absorber size, these values will be used as pixels